refactor: clean up comments and stabilize hexagonal structure for product-service
