+++ purchase-service/src/main/java/com/celotts/purchaseservice/infrastructure/adapter/input/rest/exception/GlobalExceptionHandler.java
@@ -22,7 +22,6 @@ public class GlobalExceptionHandler {
         return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
     }
 
-    // Captura errores de validación (@Valid)
     @ExceptionHandler(org.springframework.web.bind.MethodArgumentNotValidException.class)
     public ResponseEntity<Map<String, String>> handleValidation(org.springframework.web.bind.MethodArgumentNotValidException ex) {
         Map<String, String> errors = new HashMap<>();
