+++ purchase-service/src/main/java/com/celotts/purchaseservice/infrastructure/adapter/input/rest/mapper/PurchaseRestMapper.java
@@ -8,10 +8,8 @@ import org.mapstruct.Mapping;
 @Mapper(componentModel = "spring")
 public interface PurchaseRestMapper {
 
-    // Convierte lo que viene de la web al modelo de negocio
     @Mapping(target = "id", ignore = true)
     PurchaseModel toModel(PurchaseRequest request);
 
-    // Convierte el modelo de negocio a una respuesta (opcional si usas el modelo como respuesta)
     PurchaseRequest toResponse(PurchaseModel model);
 }
