+++ purchase-service/src/main/java/com/celotts/purchaseservice/infrastructure/common/validation/ValidationGroups.java
@@ -2,12 +2,9 @@ package com.celotts.purchaseservice.infrastructure.common.validation;
 
 public final class ValidationGroups {
 
-    // Constructor privado para evitar instanciación
     private ValidationGroups() {}
 
-    // Grupo de validación usado al crear entidades (POST)
     public interface Create {}
 
-    // Grupo de validación usado al actualizar entidades (PUT/PATCH)
     public interface Update {}
 }
